More twisted mass fixes
git-svn-id: http://lattice.bu.edu/qcdalg/cuda/quda@1076 be54200a-260c-0410-bdd7-ce6af2a381ab

+++ lib/tm_dslash_def.h
@@ -50,7 +50,7 @@
 #if (DD_PREC == 0)
 #define DD_PARAM4 const int oddBit, const double a, const double b
 #else
-#define DD_PARAM4 const int oddBit, const double a, const float b
+#define DD_PARAM4 const int oddBit, const float a, const float b
 #endif
 #else            // xpay
 #define DSLASH_XPAY
